dpr-assignments-2022/requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.8.2
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
filelock==3.0.12
funcsigs==1.0.2
gin-config==0.3.0
google==2.0.3
grpcio==1.27.2
idna==2.9
imageio==2.8.0
importlib-metadata==1.5.0
joblib==0.14.1
jsonschema==3.2.0
kiwisolver==1.1.0
matplotlib==3.1.3
more-itertools==8.2.0
multidict==4.7.5
networkx==2.4
numpy==1.18.1
opencv-python==4.2.0.32
packaging==20.1
pandas==1.0.1
Pillow==7.0.0
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
py-spy==0.3.3
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
redis==3.4.1
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
soupsieve==2.0
wcwidth==0.1.8
wget==3.2
yarl==1.4.2
zipp==3.0.0
